# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/accounts": {
            "post": {
                "description": "Create an account",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "account"
                ],
                "summary": "Create an account",
                "parameters": [
                    {
                        "description": "Data of the account",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.AccountRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handlers.ResponseWrapper"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/ent.Account"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ResponseWrapper"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ResponseWrapper"
                        }
                    }
                }
            }
        },
        "/accounts/{accountID}": {
            "get": {
                "description": "Retrieve an account using an account ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "account"
                ],
                "summary": "Retrieve an account using an account ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Account ID",
                        "name": "accountID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handlers.ResponseWrapper"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/ent.Account"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ResponseWrapper"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ResponseWrapper"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ResponseWrapper"
                        }
                    }
                }
            }
        },
        "/accounts/{accountID}/transactions": {
            "get": {
                "description": "Retrieve transactions of an account",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "account"
                ],
                "summary": "Retrieve transactions of an account",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Account ID",
                        "name": "accountID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "enum": [
                            "all",
                            "incoming",
                            "outgoing"
                        ],
                        "type": "string",
                        "description": "Type of transactions",
                        "name": "type",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handlers.ResponseWrapper"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/ent.Transaction"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ResponseWrapper"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ResponseWrapper"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ResponseWrapper"
                        }
                    }
                }
            }
        },
        "/accounts/{accountID}/transfer": {
            "post": {
                "description": "Transfer an amount from an account's wallet to another",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "account"
                ],
                "summary": "Transfer money",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Account ID",
                        "name": "accountID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Account and amount to transfer",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.TransferRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handlers.ResponseWrapper"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/ent.Transaction"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ResponseWrapper"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ResponseWrapper"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "ent.Account": {
            "type": "object",
            "required": [
                "balance",
                "created_at",
                "edges",
                "id",
                "name"
            ],
            "properties": {
                "balance": {
                    "description": "Balance holds the value of the \"balance\" field.",
                    "type": "number"
                },
                "created_at": {
                    "description": "CreatedAt holds the value of the \"created_at\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the AccountQuery when eager-loading is set.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/ent.AccountEdges"
                        }
                    ]
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "name": {
                    "description": "Name holds the value of the \"name\" field.",
                    "type": "string"
                }
            }
        },
        "ent.AccountEdges": {
            "type": "object",
            "required": [
                "incoming_transactions",
                "outgoing_transactions"
            ],
            "properties": {
                "incoming_transactions": {
                    "description": "IncomingTransactions holds the value of the incoming_transactions edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Transaction"
                    }
                },
                "outgoing_transactions": {
                    "description": "OutgoingTransactions holds the value of the outgoing_transactions edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Transaction"
                    }
                }
            }
        },
        "ent.Transaction": {
            "type": "object",
            "required": [
                "amount",
                "created_at",
                "edges",
                "id"
            ],
            "properties": {
                "amount": {
                    "description": "Amount holds the value of the \"amount\" field.",
                    "type": "number"
                },
                "created_at": {
                    "description": "CreatedAt holds the value of the \"created_at\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the TransactionQuery when eager-loading is set.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/ent.TransactionEdges"
                        }
                    ]
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.TransactionEdges": {
            "type": "object",
            "required": [
                "from_account",
                "to_account"
            ],
            "properties": {
                "from_account": {
                    "description": "FromAccount holds the value of the from_account edge.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/ent.Account"
                        }
                    ]
                },
                "to_account": {
                    "description": "ToAccount holds the value of the to_account edge.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/ent.Account"
                        }
                    ]
                }
            }
        },
        "handlers.AccountRequest": {
            "type": "object",
            "required": [
                "initial_balance",
                "name"
            ],
            "properties": {
                "initial_balance": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "handlers.ResponseWrapper": {
            "type": "object",
            "required": [
                "code",
                "data",
                "error"
            ],
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "error": {
                    "type": "string"
                }
            }
        },
        "handlers.TransferRequest": {
            "type": "object",
            "required": [
                "account_id",
                "amount"
            ],
            "properties": {
                "account_id": {
                    "type": "integer"
                },
                "amount": {
                    "type": "number"
                }
            }
        }
    }
}